Correction to r6160
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@6191 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/qgis.cpp
@@ -37,5 +37,5 @@ const char* QGis::qgisFeatureTypes[] =
   "WKBMultiPolygon"
 };
 
-const int QGis::DEFAULT_IDENTIFY_RADIUS=0.5;
+const double QGis::DEFAULT_IDENTIFY_RADIUS=0.5;
